backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
numpy==1.26.4
opencv-python==4.9.0.80
torch==2.2.1
torchvision==0.17.1
transformers==4.38.2
pillow==10.2.0
python-dotenv==1.0.1 
